requirements.txt
lxml>=4.5.1
torch>=1.1.0
flair==0.4.5
pandas>=1.0.5
allennlp==0.9.0
usaddress==0.5.10
beautifulsoup4>=4.9.1
usaddress-scourgify==0.2.3


